feat: update settings (#261) 탈퇴하기 컴포넌트 생성
